add Contribute link to navbar
First draft of contribution instructions
