+++ docs/sitemap.xml
@@ -2,42 +2,46 @@
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
   <url>
     <loc>https://massspec.academy/index.html</loc>
-    <lastmod>2025-03-07T15:59:09.847Z</lastmod>
+    <lastmod>2025-05-04T03:01:49.367Z</lastmod>
   </url>
   <url>
     <loc>https://massspec.academy/overview.html</loc>
-    <lastmod>2025-03-07T14:56:02.640Z</lastmod>
+    <lastmod>2025-05-04T03:01:49.370Z</lastmod>
   </url>
   <url>
     <loc>https://massspec.academy/background.html</loc>
-    <lastmod>2025-03-07T15:41:08.645Z</lastmod>
+    <lastmod>2025-05-04T03:01:49.288Z</lastmod>
   </url>
   <url>
     <loc>https://massspec.academy/ion_sources.html</loc>
-    <lastmod>2025-03-07T14:56:02.632Z</lastmod>
+    <lastmod>2025-05-04T03:01:49.369Z</lastmod>
   </url>
   <url>
     <loc>https://massspec.academy/baselines.html</loc>
-    <lastmod>2025-03-07T14:56:02.620Z</lastmod>
+    <lastmod>2025-05-04T03:01:49.288Z</lastmod>
   </url>
   <url>
     <loc>https://massspec.academy/peak_tails_abundance_sensitivity.html</loc>
-    <lastmod>2025-03-07T14:56:02.644Z</lastmod>
+    <lastmod>2025-05-04T03:01:49.371Z</lastmod>
   </url>
   <url>
     <loc>https://massspec.academy/U_Pb.html</loc>
-    <lastmod>2025-03-07T14:56:02.647Z</lastmod>
+    <lastmod>2025-05-04T03:01:49.286Z</lastmod>
   </url>
   <url>
     <loc>https://massspec.academy/basic_stats_refresher.html</loc>
-    <lastmod>2025-03-07T14:56:02.623Z</lastmod>
+    <lastmod>2025-05-04T03:01:49.289Z</lastmod>
   </url>
   <url>
     <loc>https://massspec.academy/references.html</loc>
-    <lastmod>2025-03-07T15:59:40.263Z</lastmod>
+    <lastmod>2025-05-04T03:01:49.372Z</lastmod>
   </url>
   <url>
     <loc>https://massspec.academy/Mass-Spectrometry-Book.pdf</loc>
-    <lastmod>2025-03-07T16:05:58.511Z</lastmod>
+    <lastmod>2025-05-17T03:35:43.067Z</lastmod>
+  </url>
+  <url>
+    <loc>https://massspec.academy/contributing.html</loc>
+    <lastmod>2025-05-18T14:45:05.572Z</lastmod>
   </url>
 </urlset>
